Moving files for better readability.
