+++ src/classes/vehicle/vehicle.test.ts
