+++ src/classes/vehicle/vehicle.ts
@@ -2,8 +2,8 @@
 
 import inquirer from "inquirer";
 import { GameGrid } from "../grid/grid";
-import { uniqueID } from "../utils/counter";
-import { AnyKey } from "../utils/pressanykey";
+import { uniqueID } from "../../types/utils/counter";
+import { AnyKey } from "../../types/utils/pressanykey";
 type cardinals = "N" | "S" | "E" | "W";
 
 /**
